# apps/api-go/api/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/orders/me:
    get:
      operationId: OrdersController_myOrders_v1
      parameters: []
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 我的订单
      tags:
      - orders
  /api/v1/orders/{id}:
    get:
      operationId: OrdersController_findOne_v1
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: 单订单详情
      tags:
      - orders
  /api/v1/orders:
    post:
      operationId: OrdersController_create_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 创建订单（免费商品会直接注册）
      tags:
      - orders
  /api/v1/orders/{id}/pay:
    post:
      operationId: OrdersController_pay_v1
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MockPayDto'
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 开发环境模拟支付（生产环境关闭）
      tags:
      - orders
  /api/v1/orders/{id}/cancel:
    post:
      operationId: OrdersController_cancel_v1
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 取消未支付订单
      tags:
      - orders
  /api/v1/orders/{id}/refund:
    post:
      operationId: OrdersController_refund_v1
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundOrderDto'
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: 开发环境模拟退款（生产环境关闭）
      tags:
      - orders
...
  /api/v1/enum-translations:
    get:
      operationId: CmsEnumController_listEnumTranslations_v1
      parameters:
      - name: type
        required: true
        in: query
        schema:
          type: string
      - name: locale
        required: true
        in: query
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: 枚举 i18n (label + color + icon)
      tags:
      - cms-enum
  /api/v1/date-format-templates:
    get:
      operationId: CmsEnumController_listDateFormatTemplates_v1
      parameters:
      - name: scope
        required: true
        in: query
        schema:
          type: string
      - name: locale
        required: true
        in: query
        schema:
          type: string
      responses:
        '200':
          description: ''
      summary: 日期/时间格式模板
      tags:
      - cms-enum
...
info:
  title: AI Academy API
  description: AI and LLM education platform API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateOrderDto:
      type: object
      properties:
        type:
          type: string
          enum:
          - course
          - degree
          description: 订单类型（course / degree）
        courseId:
          type: string
          description: 课程 UUID（type=course 时必填）
        degreeId:
          type: string
          description: 学位 UUID（type=degree 时必填）
        paymentMethod:
          type: string
          description: 支付方式（wechat / alipay / stripe）
      required:
      - type
    MockPayDto:
      type: object
      properties:
        paymentMethod:
          type: string
          description: 支付方式（wechat / alipay / stripe）
    RefundOrderDto:
      type: object
      properties:
        reason:
          type: string
          description: 退款原因（mock 模式不持久化）
...